deleted img paths from article dummy data and refactored styled components
