+++ server/database/article_category_data.js
@@ -61,14 +61,14 @@ const buildArticleCategory = async () => {
       const Articles = await Article.insertMany([
         {
           title: "Low mood and depression",
-          image: "Articles/image/conmunication_17-Dec-2018-10:06.jpg",
+          image: "",
           content: `But a low mood will tend to lift after a few days or weeks.
         Making some small changes in your life, such as resolving a difficult situation, talking about your problems or getting more sleep, can usually improve your mood.
-        
+
         A low mood that doesn't go away can be a sign of depression.
-        
+
         Symptoms of depression can include the following:
-        
+
         low mood lasting 2 weeks or more
         not getting any enjoyment out of life
         feeling hopeless
@@ -83,19 +83,19 @@ const buildArticleCategory = async () => {
         },
         {
           title: "Anxiety in children",
-          image: "Articles/image/Anxiety-17-Dec-2018-10:06.jpg",
-          content: `Children tend to feel anxious about different things at different ages. 
+          image: "",
+          content: `Children tend to feel anxious about different things at different ages.
         Many of these worries are a normal part of growing up.
 
-        From about eight months to three years, for example, it's very common for young children to have something called separation anxiety. 
-        They may become clingy and cry when separated from their parents or carers. 
+        From about eight months to three years, for example, it's very common for young children to have something called separation anxiety.
+        They may become clingy and cry when separated from their parents or carers.
         This is a normal stage in children's development and tends to ease off at around age two to three.
-        
-        It's also common for pre-school children to develop specific fears or phobias. 
-        Common fears in early childhood include animals, insects, storms, heights, water, blood, and the dark. 
+
+        It's also common for pre-school children to develop specific fears or phobias.
+        Common fears in early childhood include animals, insects, storms, heights, water, blood, and the dark.
         These fears usually go away gradually on their own.
-        
-        Throughout a child's life there will be other times when they feel anxiety. 
+
+        Throughout a child's life there will be other times when they feel anxiety.
         Lots of children feel anxious when going to a new school, for example, or before tests and exams.
         Some children feel shy in social situations and may need support with this.
         `,
@@ -104,13 +104,13 @@ const buildArticleCategory = async () => {
         },
         {
           title: "How to deal with stress",
-          image: "Articles/image/stress-17-Dec-2018-10:06.jpg",
+          image: "",
           content: `Stress causes physical changes in the body designed to help you take on threats or difficulties.
 
-        You may notice that your heart pounds, your breathing quickens, your muscles tense, and you start to sweat. 
+        You may notice that your heart pounds, your breathing quickens, your muscles tense, and you start to sweat.
         This is sometimes known as the fight or flight response.
-        
-        Once the threat or difficulty passes, these physical effects usually fade. 
+
+        Once the threat or difficulty passes, these physical effects usually fade.
         But if you're constantly stressed, your body stays in a state of high alert and you may develop stress-related symptoms.
         `,
           profile: therapistProfile._id,
@@ -118,16 +118,16 @@ const buildArticleCategory = async () => {
         },
         {
           title: "I used to hit my husband",
-          image: "Articles/image/hit_husband-17-Dec-2018-10:06.jpg",
+          image: "",
           content: `Intense anger caused Florence Terry to hit her husband. An anger-management course helped her regain control, and changed her life.
 
         "The first time I hit my husband was about 14 years ago. I was cross with him and lost my temper.
         He was upset and I felt dreadful, and cried and apologised. I felt scared and ashamed, but thought it was a one-off.
-        
-        "It didn't happen again for many months, maybe as many as 18 months. 
-        During that time there was verbal criticism and crossness from me, but nothing violent. I can't actually remember the second time I hit him. 
+
+        "It didn't happen again for many months, maybe as many as 18 months.
+        During that time there was verbal criticism and crossness from me, but nothing violent. I can't actually remember the second time I hit him.
         I'm sorry to say it became a pattern.
-        
+
         "Looking back, I was under a lot of stress from my job as a divorce lawyer, and I was packing my free time with other commitments, such as charity work."
         `,
           profile: therapistProfile._id,
